Fix all clippy warnings and update README for sidebar layout
Resolve all clippy warnings across workspace: derive Default for
VirtspawnConfig, remove unnecessary u32 casts in libvirt bindings,
replace redundant closures with function references in daemon routes,
use RangeInclusive::contains for validation, and restructure
compute_dashboard to avoid field reassignment. Update README keyboard
reference to document sidebar/content focus model and sub-tabs.

+++ virtspawn-core/src/libvirt/device.rs
@@ -21,7 +21,7 @@ pub fn attach_disk(conn: &Connect, vm_name: &str, req: &AttachDiskRequest) -> Re
 
     let flags = domain.get_info()
         .map(|info| {
-            if info.state as u32 == 1 {
+            if info.state == 1 {
                 virt::sys::VIR_DOMAIN_AFFECT_LIVE | virt::sys::VIR_DOMAIN_AFFECT_CONFIG
             } else {
                 virt::sys::VIR_DOMAIN_AFFECT_CONFIG
@@ -48,7 +48,7 @@ pub fn detach_disk(conn: &Connect, vm_name: &str, target: &str) -> Result<(), Li
 
     let flags = domain.get_info()
         .map(|info| {
-            if info.state as u32 == 1 {
+            if info.state == 1 {
                 virt::sys::VIR_DOMAIN_AFFECT_LIVE | virt::sys::VIR_DOMAIN_AFFECT_CONFIG
             } else {
                 virt::sys::VIR_DOMAIN_AFFECT_CONFIG
